Develop

refactor:客户模块

See merge request !75

+++ web/src/main/java/com/cardpay/controller/customer/CustomerTransferController.java
@@ -3,8 +3,8 @@ package com.cardpay.controller.customer;
 import com.cardpay.basic.base.model.ResultTo;
 import com.cardpay.basic.base.model.SelectModel;
 import com.cardpay.controller.base.BaseController;
-import com.cardpay.mgt.customer.customertransfer.model.po.TCustomerTransfer;
-import com.cardpay.mgt.customer.customertransfer.service.CustomerTransferService;
+import com.cardpay.mgt.customer.model.TCustomerTransfer;
+import com.cardpay.mgt.customer.service.CustomerTransferService;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -22,7 +22,7 @@ import java.util.List;
 @Api(value = "/customerTransfer", description = "客户移交")
 @RestController
 @RequestMapping("/customerTransfer")
-public class CustomerTransferController extends BaseController<TCustomerTransfer, Long> {
+public class CustomerTransferController extends BaseController<TCustomerTransfer, Integer> {
 
     @Autowired
     private CustomerTransferService customerTransferService;
