+++ web/src/main/java/com/cardpay/controller/manager/CustomerManagerController.java
@@ -1,7 +1,7 @@
 package com.cardpay.controller.manager;
 
 import com.cardpay.controller.base.BaseController;
-import com.cardpay.mgt.customer.customermanager.model.po.TCustomerManager;
+import com.cardpay.mgt.manager.customermanager.model.TCustomerManager;
 import io.swagger.annotations.Api;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -14,5 +14,5 @@ import org.springframework.web.bind.annotation.RestController;
 @Api(value = "/customerManager", description = "客户经理管理")
 @RestController
 @RequestMapping("/customerManager")
-public class CustomerManagerController extends BaseController<TCustomerManager, Long> {
+public class CustomerManagerController extends BaseController<TCustomerManager, Integer> {
 }
